Safari-609.1.11.1/LayoutTests/http/tests/cache/disk-cache/disk-cache-revalidation-new-expire-header.html

<script src="/js-test-resources/js-test-pre.js"></script>
<script src="resources/cache-test.js"></script>
<body>
<script>

var tests =
[
 { responseHeaders: {'Expires': 'now(0)', 'ETag': 'match' }, expiresInFutureIn304: true },
];

description("Test that the 'Expires' header is updated upon successful validation.");

debug("");

runTests(tests, function() {
    debug("304 response included an 'Expires' header in the future, so we should not need to revalidate this time.");
    runTests(tests);
});

</script>
<script src="/js-test-resources/js-test-post.js"></script>
